move Document to base
