+++ src/embed/embedding/huggingface.py
@@ -1,10 +1,10 @@
-# create embedding and store in vector db
-def huggingface_custom(document, tokenizer, model):
+# create embedding with a huggingface model
+def huggingface_custom(document, tokenizer, model, length=512):
     '''
     Create an embedding from the provided document
     '''
     for chunk in document.text:
-        inputs = tokenizer(chunk, padding=True, truncation=True, return_tensors="pt", max_length=512)
+        inputs = tokenizer(chunk, padding=True, truncation=True, return_tensors="pt", max_length=length)
         result = model(**inputs)
         embeddings = result.last_hidden_state[:, 0, :].cpu().detach().numpy()
         lst = embeddings.flatten().tolist()
