+++ src/embed/objects/__init__.py
@@ -1,5 +1,7 @@
 from embed.objects.webpage import WebPage
+from embed.objects.base import Document
 
 __all__ = [
     "WebPage",
+    "Document"
 ]
